Created the webapp
Created the webapp according to requirements.
User can search a location and the webapp will use the api to get the weather information of the location and display it.
The user's search history will be shown and its stored in the localStorage of the user's browser.
The user can also set the theme of the webapp and it is stored in the context and also then stored in the localStorage.